+++ src/Common/AppTools/ProtectedRoute.js
@@ -1,5 +1,6 @@
 import React from "react";
 import { Redirect, useHistory } from "react-router-dom";
+import "./protectedRoute.css"
 
 // You can pass props using the spread operator to throw them on an object if there are too many to break out
 const ProtectedRoute = ({ component: Component, flag, ...rest }) => {
@@ -7,15 +8,22 @@ const ProtectedRoute = ({ component: Component, flag, ...rest }) => {
     const goBackHandler = () => {
         history.goBack();
     };
-    console.log(rest); // show rest.path in the console
-    // you could redirect back to /auth if the flag is not true
+
+    const loginHandler = () => {
+        let path = '/login';
+        history.push(path);
+    };
+
     return (
         <div>
             {flag ? (
                 <Redirect to={rest.path} />
             ) : (
-                <div>
-                    <p>Unauthorized!</p> <button onClick={goBackHandler}>Go back.</button>
+                <div className="protected">
+                    <p>Unauthorized!</p>
+                    <button onClick={goBackHandler}>Go back</button>
+                    <br />
+                    <button onClick={loginHandler}>Login</button>
                 </div>
             )}
         </div>